pyproject.toml
[tool.poetry]
name = "auction"
version = "0.1.0"
description = ""
authors = ["Jofin <jofinfab@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "2.0"
utile = "1.0"
python-dotenv = "^0.17.1"
pendulum = "^2.1.2"
pycryptodome = "^3.10.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
